requirements.txt
fastapi
uvicorn
python-multipart
requests
spotipy
google-genai
yt-dlp